Remove sorting from AllowListConfig
Although a list of address is identical the same to the system regardless of
their sorting as long as the contain the same values (`["addr1", "addr2"]` is
identical as `["addr2", "addr1"]`). The first iteration of this PR would have
sorted the slice of addresses before serializing to produce a canonical slice
of bytes and the same hash

It is decided copy JSON and honour the definition from the config JSON. Most
users should copying a JSON file instead of trying to manually edit them.

At this time it is preferred to be strict about the JSON. If it is ever decided
otherwise, bring back this behaviour, this commit can be rolled back without
any side effect.