[ci] Simplify monitoring configuration
Previously deployment of prometheus and promtail in CI required bash
scripts in the repo. The latest version of the
run-monitored-tmpnet-cmd custom github action internalizes that
capability. Also:

- adds a new script - `update_avalanche_version.sh` - to ensure the
version of the custom action matches the vendored version of
avalanchego. The script is executed in the lint job to enforce this
requirement.

- adds a new script - dev_shell.sh - that starts an avalanchego dev
shell that has access to dependencies required by tmpnet for thing
like starting collectors of logs and metrics