Use explicit subscriptions instead of toEngine channel
This commit removes the toEngine channel pattern from the VM and adds an additional API:

```
WaitForEvent(ctx context.Context)(Message, error)
```

Which returns when an event (pending transactions or state sync done Message) is sent by the VM.

Unlike the toEngine pattern which only notifies the consumer a single time until a block is built, successive calls to `WaitForEvent ` will return that transactions are pending as long as the mem-pool is not empty and a block can be built, otherwise the calls block.

Having a dedicated API allows consensus to obtain information about whether a block should be built on demand and therefore correlated to a specific round, while using the current toEngine pattern makes it impossible to correlate with a specific round, and this makes it impossible to incorporate with consensus.

The EVM checks if transactions are in the mempool or waits until they arrive.

Unit tests and e2e tests

No

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>